fixes
* fix leaderboard
* switch xp value type from long to int (saves us some storage, since itll take 90 years of active time to reach max value for an int anyway)

+++ src/net/celestialgaze/GuraBot/commands/Commands.java
@@ -113,7 +113,7 @@ public class Commands {
 							}
 							
 							Document rolesDoc = sdbSettings.get("roles", new Document());
-							long currentXP = si.getXP(userId, xpDoc);
+							int currentXP = si.getXP(userId, xpDoc);
 							int random = 20+new Random().nextInt(5);
 							List<String> badRoles = new ArrayList<String>();
 							if (XPUtil.getLevel(currentXP) < XPUtil.getLevel(currentXP + random)) {
